+++ src/base/io/ResourceResolver.ts
@@ -6,6 +6,20 @@
  */
 export interface ResourceResolver {
   /**
+   * Returns the URI that results from resolving the given
+   * URI against the base URI of this resource resolver.
+   *
+   * The result will be a relative URI that includes path
+   * components that reflect the paths that have been given
+   * when this resource resolver was created by calling
+   * `derive` on another resource resolver.
+   *
+   * @param uri - The URI
+   * @returns The resolved URI
+   */
+  resolveUri(uri: string): string;
+
+  /**
    * Resolve the data from the given URI.
    *
    * The given URI may be relative to the base URI for
@@ -13,9 +27,9 @@ export interface ResourceResolver {
    *
    * @param uri - The URI
    * @returns A promise that resolves with the buffer data,
-   * or with `null` if the resource could not be resolved.
+   * or with `undefined` if the resource could not be resolved.
    */
-  resolveData(uri: string): Promise<Buffer | null>;
+  resolveData(uri: string): Promise<Buffer | undefined>;
 
   /**
    * Resolve parts of the data from the given URI.
@@ -23,7 +37,7 @@ export interface ResourceResolver {
    * The given URI may be relative to the base URI for
    * which this instance has been created.
    *
-   * If the resource cannot be resolved, then `null` is
+   * If the resource cannot be resolved, then `undefined` is
    * returned.
    *
    * Otherwise, the returned buffer contains _at least_ the
@@ -32,9 +46,12 @@ export interface ResourceResolver {
    *
    * @param uri - The URI
    * @returns A promise that resolves with the buffer data,
-   * or with `null` if the resource could not be resolved.
+   * or with `undefined` if the resource could not be resolved.
    */
-  resolveDataPartial(uri: string, maxBytes: number): Promise<Buffer | null>;
+  resolveDataPartial(
+    uri: string,
+    maxBytes: number
+  ): Promise<Buffer | undefined>;
 
   /**
    * Derive an instance from this one, with a different base